Enfiinnn ajouts des permissions pour les utilsations fonctionelle!!!!!

+++ Backend/app/Http/Controllers/ActiviteController.php
@@ -11,23 +11,23 @@ use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Controllers\Controller;
+use App\Policies\ActivityPolicy;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
+use Illuminate\Auth\Access\AuthorizationException;
 
-class ActiviteController extends Controller implements HasMiddleware
+class ActiviteController extends Controller 
 {
-    public static function middleware()
-    {
-        return [
-            new Middleware('auth:sanctum', except : ['getAllActivities', 'getActivityByDayTime', 
-            'getActivitybyType', 'getActivityBySeason',
-            'getActivityByName'])
-        ];
-    }
+   
+
+    
     protected $userService;
 
     public function __construct(ActiviteService $userService)
     {
         $this->userService = $userService;
+         $this->middleware('auth:sanctum');
     }
 
     public function getAllActivities() {
@@ -35,7 +35,12 @@ class ActiviteController extends Controller implements HasMiddleware
     }
 
     public function modifyActivity(int $activiteId, Request $request) {
-         try {
+         
+          
+             $activite = Activite::findOrFail($activiteId);
+             $this->authorize('update', $activite);
+
+        
             $validatedInputActivity = $request->validate([
                 'titre' => 'required|string|max:255',
                 'description' => 'required|string',
@@ -46,14 +51,18 @@ class ActiviteController extends Controller implements HasMiddleware
 
             $userValidated = $this->userService->updateActiviy($activiteId, $validatedInputActivity);
             return response()->json($userValidated, 202);
-        } catch (\Exception $e) {
+    
             return response()->json($e->getMessage(),500);
-        }
+        
     }
 
    public function deleteActivityById(int $activiteId)
     {
         try {
+
+            $activite = Activite::findOrFail($activiteId);
+           
+            Gate::authorize('delete', $activite);     
             $this->userService->deleteActivity($activiteId);
         } catch (ModelNotFoundException $e) {
         return response()->json(['error' => "Activity $activiteId not found"], 404);
@@ -72,7 +81,7 @@ class ActiviteController extends Controller implements HasMiddleware
 
         $user =  $contenu->user();
         $userComment =$this->userService->addCommentToActivityFromUser( $user->id , $activityId, $validated['contenu']);
-        
+
         return response()->json($userComment);
     } catch (\Exception $e) {
         return response()->json($e->getMessage());
@@ -125,6 +134,9 @@ class ActiviteController extends Controller implements HasMiddleware
         $activiteTitle = $this->userService->getActivitiesByName($validated['title']);
         return response()->json($activiteTitle);
     }
+
+
+
